Implement required code changes for common target to compile as C++
* Add casts where necessary mostly for void * -> any other type
* Replaced use of min, max, swap macros with std variant
* Fixed where `const char` strings were passed/stored to non-const variables
* Updated VECTOR_RESIZE with C++ variant that relays on decltype to make type safe cast from memory manager

+++ src/common/sql.c
@@ -561,7 +561,7 @@ static int Sql_P_BindSqlDataType(MYSQL_BIND *bind, enum SqlDataType buffer_type,
 		break;
 	default:
 	case SQLDT_LASTID:
-		ShowDebug("Sql_P_BindSqlDataType: unsupported buffer type (%u)\n", buffer_type);
+		ShowDebug("Sql_P_BindSqlDataType: unsupported buffer type (%u)\n", (unsigned int)buffer_type);
 		return SQL_ERROR;
 	}
 	bind->buffer = buffer;
@@ -583,7 +583,7 @@ static void Sql_P_ShowDebugMysqlFieldInfo(const char *prefix, enum enum_field_ty
 	PRAGMA_GCC46(GCC diagnostic ignored "-Wswitch-enum")
 	switch (type) {
 		default:
-			ShowDebug("%stype=%s%u, length=%lu\n", prefix, sign, type, length);
+			ShowDebug("%stype=%s%u, length=%lu\n", prefix, sign, (unsigned int)type, length);
 			return;
 #define SHOW_DEBUG_OF(x) case x: type_string = #x; break
 		SHOW_DEBUG_OF(MYSQL_TYPE_TINY);
@@ -814,7 +814,7 @@ static int SqlStmt_BindColumn(struct SqlStmt *self, size_t idx, enum SqlDataType
 
 	if (buffer_type == SQLDT_STRING || buffer_type == SQLDT_ENUM) {
 		if (buffer_len < 1) {
-			ShowDebug("SqlStmt_BindColumn: buffer_len(%"PRIuS") is too small, no room for the null-terminator\n", buffer_len);
+			ShowDebug("SqlStmt_BindColumn: buffer_len(%" PRIuS ") is too small, no room for the null-terminator\n", buffer_len);
 			return SQL_ERROR;
 		}
 		--buffer_len;// null-terminator
@@ -1074,7 +1074,7 @@ void Sql_HerculesUpdateCheck(struct Sql *self)
 			if( SQL_ERROR == SQL->Query(self, "SELECT 1 FROM `sql_updates` WHERE `timestamp` = '%u' LIMIT 1", timestampui) )
 				Sql_ShowDebug(self);
 			if( Sql_NumRows(self) != 1 ) {
-				StrBuf->Printf(&buf,CL_MAGENTA"[SQL]"CL_RESET": -- '"CL_WHITE"%s"CL_RESET"'\n", path);
+				StrBuf->Printf(&buf,CL_MAGENTA "[SQL]" CL_RESET ": -- '" CL_WHITE "%s" CL_RESET "'\n", path);
 				performed++;
 			}
 		}
@@ -1085,7 +1085,7 @@ void Sql_HerculesUpdateCheck(struct Sql *self)
 	fclose(ifp);
 
 	if( performed ) {
-		ShowSQL("- detected %u new "CL_WHITE"SQL updates"CL_RESET"\n",performed);
+		ShowSQL("- detected %u new " CL_WHITE "SQL updates" CL_RESET "\n",performed);
 		ShowMessage("%s",StrBuf->Value(&buf));
 		ShowSQL("To manually skip, type: 'sql update skip <file name>'\n");
 	}
